fix(merge): suppression des fichiers provenant de la mauvaise branche

+++ project/back/src/main/java/com/puericulture/forwardtrading/repository/ChildrenRepository.java
@@ -3,4 +3,4 @@ package com.puericulture.forwardtrading.repository;
 import com.puericulture.forwardtrading.entity.ChildrenEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ChildrenRepository extends CrudRepository<ChildrenEntity, Long> {}
+public interface ChildrenRepository extends CrudRepository<ChildrenEntity, Integer> {}
